Renamed math.h to mathUtils.h
