+++ source/CaveStoryEngine.vcxproj.filters
@@ -57,9 +57,6 @@
     <ClCompile Include="level.cpp">
       <Filter>Source Files\modules</Filter>
     </ClCompile>
-    <ClCompile Include="math.cpp">
-      <Filter>Source Files\modules</Filter>
-    </ClCompile>
     <ClCompile Include="filesystem.cpp">
       <Filter>Source Files\modules</Filter>
     </ClCompile>
@@ -162,15 +159,15 @@
     <ClCompile Include="fade.cpp">
       <Filter>Source Files\modules</Filter>
     </ClCompile>
+    <ClCompile Include="mathUtils.cpp">
+      <Filter>Source Files\modules</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="render.h">
       <Filter>Source Files\modules</Filter>
     </ClInclude>
     <ClInclude Include="level.h">
-      <Filter>Source Files\modules</Filter>
-    </ClInclude>
-    <ClInclude Include="math.h">
       <Filter>Source Files\modules</Filter>
     </ClInclude>
     <ClInclude Include="filesystem.h">
@@ -279,6 +276,9 @@
       <Filter>Source Files\classes\npc\act</Filter>
     </ClInclude>
     <ClInclude Include="fade.h">
+      <Filter>Source Files\modules</Filter>
+    </ClInclude>
+    <ClInclude Include="mathUtils.h">
       <Filter>Source Files\modules</Filter>
     </ClInclude>
   </ItemGroup>
